PETL-135 - Code review with Claude around performance and memory leaks (#41)
Code review fixes: resource leaks, thread-safety, and performance improvements

+++ src/test/java/org/pih/petl/job/BasePetlTest.java
@@ -83,7 +83,7 @@ public abstract class BasePetlTest {
         }
     }
 
-    public void verifyRowCount(String table, int expectedRows) throws Exception {
+    public void verifyRowCount(String table, long expectedRows) throws Exception {
         Assert.assertEquals(expectedRows, getSqlServerDatasource().rowCount(table));
     }
